WW-1421 Move struts.properties settings to struts.xml WW-1458 Replace OGNL scriptlets with key attribute
Interim commit for discussion purposes. I'm trying to bring the blank up to date with the latest practice. As to the struts.properties, I tried moving these to struts.xml, but it doesn't seem to be working. I copied the elements from the Spring plugin struts-plugin.xml, but IDEA complains that the constant element is not defined. Strangely, I don't get the same prompt for the struts-plugin.xml. I tried deploying the application, and it seems that the struts.xml constants are ignored. 

git-svn-id: https://svn.apache.org/repos/asf/struts/struts2/trunk@477484 13f79535-47bb-0310-9956-ffa450edef68