WW-1520 Add Autocompleter AJAX tag - Apply patch submitted by Musachy Barroso
In the showcase, the ajax/autocompleter/index.jsp page doesn't seem to work correctly for me. There are four dropdown controls, but only the second (using a local list) presents a list for me. The third control (force valid options) does seem to clear whatever I put in, but I don't know what the valid values are suppose to be (become no list drops down). I seem to have the same problems with IE and FF. 

git-svn-id: https://svn.apache.org/repos/asf/struts/struts2/trunk@479236 13f79535-47bb-0310-9956-ffa450edef68