[PATCH] First attempt at enabling value-labels to be pulled from  Locale Files.  Adding listValueKey attribute to checkboxList, radioMap, and  select tags.
Issues:
This implementation uses the <@s.text> tag to handle the localization with an assignment. I presume this is less than idea, but I lack the knowledge of the Struts2 tag model that would suggest where to check for this on the valueStack or elsewhere.

Example:

public enum yesno {
YES, NO;
}

<@s.radio name="yes no" list="#{yesno.values()}" valueKey="name()" />

<!-- this will look for the name() in the locale file and emit the localized value for the label of the radio button.