Fix behaviour for convertValue with value null
I insinuate that the non primitive BigInteger and BigDecimal types
should also return default values for null input.

Rename the map to `baseTypeDefaults for better understanding and
remove the check for `isPrimitive` b/c this seemed to be kind of optimization
but breaked the check for included non-primitive types.

B/c the unit tests run successfully, we can use the existing default fields ZERO.